renamed sct as sct_resistant, added fgt, background elaboration is in a separate file
